feat(develop): split plural util, fix storybook command
